+++ CHANGELOG.md
@@ -78,6 +78,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove alias `GraphQL\Validator\Rules\AbstractValidationRule`, use `GraphQL\Validator\Rules\ValidationRule`
 - Remove `OperationParams` method `getOriginalInput()` in favor of public property `$originalInput`
 - Remove `OperationParams` method `isReadOnly()` in favor of public property `$readOnly`
+- Remove `Utils::withErrorHandling()`
+- Remove `TypeComparators::doTypesOverlap()`
 
 ## 14.11.3